feat:test

+++ src/main/kotlin/com/example/object/domain/Screening.kt
@@ -13,7 +13,7 @@ class Screening(
     }
 
     private fun calculateFee(audienceCount: Int): Money {
-        return movie.calculateMovieFee(this).times(audienceCount.toDouble())
+        return movie.calculateMovieFee(this).times(audienceCount.toLong())
     }
 
     fun getMovieFee(): Money {
